pyproject.toml
[tool.poetry]
name = "diffusion"
version = "0.1.0"
description = ""
authors = ["John <john.loverich@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
denoising-diffusion-pytorch = "^0.14.3"
torch = "^1.11.0"
matplotlib = "^3.5.2"
pytorch-lightning = "^1.6.3"
protobuf = "3.20.0"
hydra-nevergrad-sweeper = "^1.2.0"
hydra-core = "^1.2.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}
jupyter = "^1.0.0"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
